moved code from provider to service
