+++ src/service/stoppoint.js
@@ -1,11 +1,45 @@
 import * as mobiliteit from '../source/stoppoint/mobiliteit';
+import fuzzy    from 'fuzzy';
+import config   from '../config';
+import distance from '../helper/distance';
+import inbox    from '../helper/inbox';
+import cron     from 'node-cron';
+
+var fuzzyOptions = {
+    extract: function(obj) { return obj.properties.name; }
+};
+
+var stopPoints = [];
+
+cron.schedule(config('MOBILITEIT_REFRESH_CRON', true), function(){
+    loadStoppoints();
+});
+
+const loadStoppoints = async () => {
+    stopPoints = await mobiliteit.load();
+};
+
+const cache = async () => {
+    if (stopPoints.length === 0) {
+        await loadStoppoints();
+    }
+};
 
 export const all = async () => {
-    return await mobiliteit.all();
+    await cache();
+    return {
+        type: 'FeatureCollection',
+        features: stopPoints
+    };
 };
 
 export const get = async stopPoint => {
-    return await mobiliteit.get(stopPoint);
+    await cache();
+    for (var i = 0; i < stopPoints.length; i++) {
+        if (stopPoints[i].properties.id == stopPoint) {
+            return stopPoints[i];
+        }
+    }
 };
 
 export const departures = async stopPoint => {
@@ -13,13 +47,60 @@ export const departures = async stopPoint => {
 };
 
 export const around = async (lon, lat, radius) => {
-    return await mobiliteit.around(lon, lat, radius);
+    await cache();
+    var dist = 0;
+    var stopPointsAround = [];
+
+    for (var i = 0; i < stopPoints.length; i++) {
+        dist = distance(
+            parseFloat(lon),
+            parseFloat(lat),
+            stopPoints[i].geometry.coordinates[0],
+            stopPoints[i].geometry.coordinates[1],
+        );
+
+        if (dist <= radius) {
+            //TODO: fix this piece of code...
+            stopPointsAround.push({
+                type: stopPoints[i].type,
+                geometry: stopPoints[i].geometry,
+                properties: {
+                    id: stopPoints[i].properties.id,
+                    name: stopPoints[i].properties.name,
+                    distance: parseFloat(dist.toFixed(2))
+                }
+            });
+        }
+    }
+    return {
+        type: 'FeatureCollection',
+        features: stopPointsAround
+    };
 };
 
 export const box = async (swlon, swlat, nelon, nelat) => {
-    return await mobiliteit.box(swlon, swlat, nelon, nelat);
+    await cache();
+    var stopPointsInBox = stopPoints.filter(function(stopPoint) {
+        return inbox(
+            swlon, swlat, nelon, nelat,
+            stopPoint.geometry.coordinates[0],
+            stopPoint.geometry.coordinates[1],
+        );
+    });
+    return {
+        type: 'FeatureCollection',
+        features: stopPointsInBox
+    };
 };
 
 export const search = async searchString => {
-    return await mobiliteit.search(searchString);
+    await cache();
+
+    var results = fuzzy.filter(searchString, stopPoints, fuzzyOptions);
+    var stopPointMatches = results.map(function(res) { return res.original; });
+
+    return {
+        type: 'FeatureCollection',
+        features: stopPointMatches
+    };
 };
