+++ src/source/stoppoint/mobiliteit.js
@@ -1,25 +1,8 @@
 import request  from 'request-promise-native';
-import fuzzy    from 'fuzzy';
 import config   from '../../config';
-import distance from '../../helper/distance';
-import inbox    from '../../helper/inbox';
-var cron = require('node-cron');
-
-var stopPoints = [];
-var fuzzyOptions = {
-    extract: function(obj) { return obj.properties.name; }
-};
 
 const getRaw = async () => {
     return await request(config('MOBILITEIT_STOPPOINTS', true));
-};
-
-cron.schedule(config('MOBILITEIT_REFRESH_CRON', true), function(){
-    loadStoppoints();
-});
-
-const loadStoppoints = async () => {
-    stopPoints = await load();
 };
 
 export const load = async () => {
@@ -50,29 +33,6 @@ export const load = async () => {
         });
     }
     return newStopPoints;
-};
-
-const cache = async () => {
-    if (stopPoints.length === 0) {
-        await loadStoppoints();
-    }
-};
-
-export const all = async () => {
-    await cache();
-    return {
-        type: 'FeatureCollection',
-        features: stopPoints
-    };
-};
-
-export const get = async stopPoint => {
-    await cache();
-    for (var i = 0; i < stopPoints.length; i++) {
-        if (stopPoints[i].properties.id == stopPoint) {
-            return stopPoints[i];
-        }
-    }
 };
 
 export const departures = async stopPoint => {
@@ -112,64 +72,4 @@ export const departures = async stopPoint => {
         }
     }
     return departures;
-};
-
-export const around = async (lon, lat, radius) => {
-    await cache();
-    var dist = 0;
-    var stopPointsAround = [];
-
-    for (var i = 0; i < stopPoints.length; i++) {
-        dist = distance(
-            parseFloat(lon),
-            parseFloat(lat),
-            stopPoints[i].geometry.coordinates[0],
-            stopPoints[i].geometry.coordinates[1],
-        );
-
-        if (dist <= radius) {
-            //TODO: fix this piece of code...
-            stopPointsAround.push({
-                type: stopPoints[i].type,
-                geometry: stopPoints[i].geometry,
-                properties: {
-                    id: stopPoints[i].properties.id,
-                    name: stopPoints[i].properties.name,
-                    distance: parseFloat(dist.toFixed(2))
-                }
-            });
-        }
-    }
-    return {
-        type: 'FeatureCollection',
-        features: stopPointsAround
-    };
-};
-
-export const box = async (swlon, swlat, nelon, nelat) => {
-    await cache();
-    var stopPointsInBox = stopPoints.filter(function(stopPoint) {
-        return inbox(
-            swlon, swlat, nelon, nelat,
-            stopPoint.geometry.coordinates[0],
-            stopPoint.geometry.coordinates[1],
-        );
-    });
-    return {
-        type: 'FeatureCollection',
-        features: stopPointsInBox
-    };
-};
-
-export const search = async searchString => {
-    await cache();
-
-    var results = fuzzy.filter(searchString, stopPoints, fuzzyOptions);
-    var stopPointMatches = results.map(function(res) { return res.original; });
-
-    return {
-        type: 'FeatureCollection',
-        features: stopPointMatches
-    };
-
 };
